Cleaned up and single "happy path spec".
* Lexer, Parser and APIs in different files.
* A single test which parses a simple valid input and outputs a CST.
