+++ lib/dsl/poc/lexer.js
@@ -3,7 +3,6 @@ const _ = require('lodash');
 
 
 const Lexer = chevrotain.Lexer;
-const Parser = chevrotain.Parser;
 
 // ----------------- lexer -----------------
 const tokens = {};
@@ -146,120 +145,7 @@ createToken({ name: 'DOT', pattern: '.' });
 // It is useful to help debug the token vector results.
 const JDLLexer = new Lexer(_.values(tokens), { debug: true });
 
-// short prefix to reduce verbosity.
-const t = tokens;
-
-class JDLParser extends Parser {
-  constructor(input) {
-    super(input, tokens);
-
-    const $ = this;
-
-    // HIGHLIGHTS1: Any rule may be used as a start rule, there is no artificial limit
-    // like in pegjs. This capability is useful for partial parsing, e.g.:
-    // 1. Code snippets
-    // 2. Incremental parsing of only the changed parts of an active Editor.
-    // 3. writing Unit tests for micro code samples.
-    $.RULE('prog', () => {
-      $.OR([
-        { ALT: () => { $.SUBRULE($.constantDecl); } },
-        { ALT: () => { $.SUBRULE($.entityDecl); } }
-      ]);
-    });
-
-    $.RULE('constantDecl', () => {
-      $.CONSUME(t.NAME);
-      $.CONSUME(t.EQUALS);
-      $.CONSUME(t.INTEGER);
-    });
-
-    $.RULE('entityDecl', () => {
-      $.CONSUME(t.ENTITY);
-      $.CONSUME(t.NAME);
-
-      $.OPTION(() => {
-        $.SUBRULE($.entityTableNameDecl);
-      });
-
-      // the "2" suffix is a quirk of Chevrotain, more details:
-      // https://github.com/SAP/chevrotain/blob/master/docs/faq.md#-why-are-the-unique-numerical-suffixes-consume1consume2-needed-for-the-dsl-rules
-      $.OPTION2(() => {
-        $.SUBRULE($.entityBody);
-      });
-    });
-
-    $.RULE('entityTableNameDecl', () => {
-      $.CONSUME(t.LPAREN);
-      $.CONSUME(t.NAME);
-      $.CONSUME(t.RPAREN);
-    });
-
-    $.RULE('entityBody', () => {
-      $.CONSUME(t.LCURLY);
-      $.AT_LEAST_ONE(() => {
-        $.SUBRULE($.fieldDec);
-      });
-      $.CONSUME(t.RCURLY);
-    });
-
-    $.RULE('fieldDec', () => {
-      $.CONSUME(t.NAME);
-      $.SUBRULE($.type);
-      // Short form for: "(X(,X)*)?"
-      $.MANY_SEP({
-        SEP: t.COMMA,
-        DEF: () => {
-          $.SUBRULE($.validation);
-        }
-      });
-      $.CONSUME(t.RCURLY);
-    });
-
-    $.RULE('type', () => {
-      $.CONSUME($.NAME);
-    });
-
-    $.RULE('validation', () => {
-      $.OR([
-        { ALT: () => { $.CONSUME(t.REQUIRED); } },
-        { ALT: () => { $.SUBRULE($.minMaxValidation); } },
-        { ALT: () => { $.SUBRULE($.pattern); } }
-      ]);
-    });
-
-    $.RULE('minMaxValidation', () => {
-      // HIGHLIGHT:
-      // Note that "MIN_MAX_KEYWORD" is an abstract token and could match 6 different concrete token types
-      $.CONSUME(t.MIN_MAX_KEYWORD);
-      $.CONSUME(t.LPAREN);
-      $.OR([
-        { ALT: () => { $.CONSUME(t.INTEGER); } },
-        { ALT: () => { $.CONSUME(t.NAME); } }
-      ]);
-      $.CONSUME(t.RPAREN);
-    });
-
-    $.RULE('pattern', () => {
-      $.CONSUME(t.PATTERN);
-      $.CONSUME(t.LPAREN);
-      // HIGHLIGHT:
-      // With Chevrotain the grammar can be debugged directly by using good old fashioned breakpoints.
-      // No need to to try and figure out a 10,000 lines generated file, or worse not even have that
-      // if we would be use some JS combinator.
-      // debugger; 
-      $.CONSUME(t.REGEX);
-      $.CONSUME(t.RPAREN);
-    });
-
-    // very important to call this after all the rules have been defined.
-    // otherwise the parser may not work correctly as it will lack information
-    // derived during the self analysis phase.
-    Parser.performSelfAnalysis(this);
-  }
-}
-
 module.exports = {
   tokens,
-  JDLLexer,
-  JDLParser
+  JDLLexer
 };
